Fix lowOrHi class value on line 51
Interestingly, this error was easy to get ignored. We were trying to set the value of the const lowOrHi with the CSS class element .lowOrHi. But without using the dot in front of it. No syntax error was thrown, but since the querySelector didn't provide any value, the constant's value remained null. Fascinating.