Fix random number formula
Since Math.random give values between 0 (inclusive) and 1 (exclusive), the fundamental problem arises when we use Math.floor on it. Because Math.floor rounds up the number to the nearest whole number and for any value of Math.random, it is going to be 0. When we add 1 to this zero, it is no wonder that the value used to be 1 always.

In this exercise, I also learned to use console.log(randomNumber) right after defining randomNumber's value. And it showed how we can check the value and make sure everything works fine.