Add cloned gist about chrome dev tools commands
I came across a Dev Tools video showcasing differnt console commands to help us set up tracking errors and execution in the code and browser console. Amazing video, this is the clone of the files used in that video. I will be implementing some of these codes in the number-bame-errors.html file