Modify the game to enable five rounds of play
As instructed in the requirement on the assignment page, moved all the code to a new function called playGame. Also wrapped up the necessary code to for loop. The for loop runs for 5 times. Allowing to play 5 rounds of game.

I also added many console.log statements meaningfully with some separator and cleaning statements to make the game in console cleaner than usual.

Future upgrade: I want to modify the playRound function in such a way that I don't have to hardcode the code comparing strings. This would require me to assign weightage to each of the three choices and weighing them or their sum. Something like that. Also, another improvement that I will make immediately after this commit is about handling any other input from humans.